:bath: More clean up
