+++ cypress/integration/application/index.js
@@ -1,11 +1,14 @@
-import config from '../../../src/app/lib/config/services';
-import { describeForLocalOnly } from '../../support/limitEnvRuns';
+import config from '../../support/config/services';
 
-describeForLocalOnly('Application', () => {
+describe('Application', () => {
   Object.keys(config)
-    .filter(service => service !== 'default')
+    .filter(service => service !== 'news')
+    .filter(service =>
+      Object.keys(config[service].pageTypes).some(
+        pageType => config[service].pageTypes[pageType] !== undefined,
+      ),
+    )
     .forEach(service => {
-      // All services test sws
       it(`should return a 200 status code for ${service}'s service worker`, () => {
         cy.testResponseCodeAndType(
           `/${service}/sw.js`,
@@ -31,16 +34,5 @@ describe('Application', () => {
       200,
       'application/javascript',
     );
-  });
-
-  // Once all manifest are done this should be move into the object forEach above
-  ['igbo', 'news/articles', 'pidgin', 'yoruba'].forEach(service => {
-    it(`should return a 200 status code for ${service} manifest file`, () => {
-      cy.testResponseCodeAndType(
-        `/${service}/manifest.json`,
-        200,
-        'application/json',
-      );
-    });
   });
 });
