+++ cypress/integration/pages/articles.js
@@ -1,9 +1,8 @@
-import { BBC_BLOCKS } from '@bbc/psammead-assets/svgs';
+  import { BBC_BLOCKS } from '@bbc/psammead-assets/svgs';
 import iterator from '../../support/iterator';
 import envConfig from '../../support/config/envs';
 import config from '../../support/config/services';
 import appConfig from '../../../src/app/lib/config/services';
-import describeForEuOnly from '../../support/describeForEuOnly';
 import { getBlockByType, getBlockData } from '../../support/bodyTestHelper';
 
 // TODO: Remove after https://github.com/bbc/simorgh/issues/2959
@@ -15,25 +14,26 @@ const serviceHasCorrectlyRenderedParagraphs = service => service !== 'sinhala';
 
 const tests = service =>
   describe(`Tests`, () => {
-    describe(`Meta Tests`, () => {
+    describe(`Metadata`, () => {
       it('should have a correct robot meta tag', () => {
         cy.checkMetadataContent('head meta[name="robots"]', 'noodp,noydir');
       });
 
-      it('should have resource hints', () => {
-        const resources = [
-          envConfig.assetOrigin,
-          'https://ichef.bbci.co.uk',
-          'https://gel.files.bbci.co.uk',
-        ];
+      it('should include the canonical URL', () => {
+        cy.checkCanonicalURL(
+          `https://www.bbc.com${config[service].pageTypes.articles}`,
+        );
+      });
 
-        resources.forEach(resource => {
-          const selector = `head link[href="${resource}"]`;
-          cy.get(selector).should('have.attr', 'rel', 'preconnect');
-          cy.get(selector)
-            .eq(1)
-            .should('have.attr', 'rel', 'dns-prefetch');
-        });
+      it('should have lang and dir attributes', () => {
+        cy.request(`${config[service].pageTypes.articles}.json`).then(
+          ({ body }) => {
+            cy.hasHtmlLangDirAttributes({
+              lang: body.metadata.passport.language,
+              dir: appConfig[service].dir,
+            });
+          },
+        );
       });
 
       it('should have the correct facebook metadata', () => {
@@ -75,122 +75,46 @@ const tests = service =>
           .and('contain', 'author')
           .and('contain', 'headline');
       });
-    });
 
-    it('Metadata', () => {
-      cy.request(`${config[service].pageTypes.articles}.json`).then(
-        ({ body }) => {
-          cy.get('meta[name="description"]').should(
-            'have.attr',
-            'content',
-            body.promo.summary || body.promo.headlines.seoHeadline,
-          );
-          cy.get('meta[name="og:title"]').should(
-            'have.attr',
-            'content',
-            body.promo.headlines.seoHeadline,
-          );
-          cy.get('meta[name="og:type"]').should(
-            'have.attr',
-            'content',
-            body.metadata.type,
-          );
-          cy.get('meta[name="article:published_time"]').should(
-            'have.attr',
-            'content',
-            new Date(body.metadata.firstPublished).toISOString(),
-          );
-          cy.get('meta[name="article:modified_time"]').should(
-            'have.attr',
-            'content',
-            new Date(body.metadata.lastPublished).toISOString(),
-          );
-          cy.get('html').should(
-            'have.attr',
-            'lang',
-            body.metadata.passport.language,
-          );
-        },
-      );
-
-      it('should include mainEntityOfPage in the LinkedData', () => {
-        cy.get('script[type="application/ld+json"]')
-          .should('contain', 'mainEntityOfPage')
-          .and('contain', 'author')
-          .and('contain', 'headline');
-      });
-    });
-
-    describeForEuOnly('Consent Banners', () => {
-      it('have correct translations', () => {
-        cy.hasConsentBannerTranslations(service);
-      });
-    });
-
-    it('should have lang and dir attributes', () => {
-      cy.request(`${config[service].pageTypes.articles}.json`).then(
-        ({ body }) => {
-          cy.hasHtmlLangDirAttributes({
-            lang: body.metadata.passport.language,
-            dir: appConfig[service].dir,
-          });
-        },
-      );
-    });
-
-    describe('Footer Tests', () => {
-      it('should render the BBC News branding', () => {
-        cy.get('footer a')
-          .eq(0)
-          .should(
-            'contain',
-            appConfig[service].serviceLocalizedName !== undefined
-              ? `${appConfig[service].product}, ${appConfig[service].serviceLocalizedName}`
-              : appConfig[service].product,
-          );
-      });
-
-      it('should have working links', () => {
-        cy.get('footer ul').within(() =>
-          appConfig[service].footer.links.forEach(({ href }, key) =>
-            cy.checkLinks(key, href),
-          ),
+      it('should include correct metadata', () => {
+        cy.request(`${config[service].pageTypes.articles}.json`).then(
+          ({ body }) => {
+            cy.get('meta[name="description"]').should(
+              'have.attr',
+              'content',
+              body.promo.summary || body.promo.headlines.seoHeadline,
+            );
+            cy.get('meta[name="og:title"]').should(
+              'have.attr',
+              'content',
+              body.promo.headlines.seoHeadline,
+            );
+            cy.get('meta[name="og:type"]').should(
+              'have.attr',
+              'content',
+              body.metadata.type,
+            );
+            cy.get('meta[name="article:published_time"]').should(
+              'have.attr',
+              'content',
+              new Date(body.metadata.firstPublished).toISOString(),
+            );
+            cy.get('meta[name="article:modified_time"]').should(
+              'have.attr',
+              'content',
+              new Date(body.metadata.lastPublished).toISOString(),
+            );
+            cy.get('html').should(
+              'have.attr',
+              'lang',
+              body.metadata.passport.language,
+            );
+          },
         );
       });
-
-      it('should contain copyright text', () => {
-        cy.get('footer p').should(
-          'contain',
-          `© ${new Date().getFullYear()} ${
-            appConfig[service].footer.copyrightText
-          }`,
-        );
-      });
-
-      it('should contain a link in the copyright text', () => {
-        cy.get('footer p')
-          .children('a')
-          .should('have.attr', 'href')
-          .and('contain', appConfig[service].footer.externalLink.href);
-      });
-    });
-
-    it('should include the canonical URL', () => {
-      cy.checkCanonicalURL(
-        `https://www.bbc.com${config[service].pageTypes.articles}`,
-      );
     });
 
     describe(`Article Body`, () => {
-      it('should render the BBC News branding', () => {
-        cy.get('header a').should(
-          'contain',
-          appConfig[service].serviceLocalizedName !== undefined
-            ? `${appConfig[service].product}, ${appConfig[service].serviceLocalizedName}`
-            : appConfig[service].product,
-        );
-      });
-
       it('should render a H1, which contains/displays a styled headline', () => {
         cy.request(`${config[service].pageTypes.articles}.json`).then(
           ({ body }) => {
