+++ cypress/support/alwaysTests.js
@@ -1,0 +1,105 @@
+import envConfig from './config/envs';
+import appConfig from '../../src/app/lib/config/services';
+import describeForEuOnly from './describeForEuOnly';
+
+const alwaysTests = service => {
+  describe('Always tests', () => {
+    describe(`Metadata`, () => {
+      it('should have resource hints', () => {
+        const resources = [
+          envConfig.assetOrigin,
+          'https://ichef.bbci.co.uk',
+          'https://gel.files.bbci.co.uk',
+        ];
+
+        resources.forEach(resource => {
+          const selector = `head link[href="${resource}"]`;
+          cy.get(selector).should('have.attr', 'rel', 'preconnect');
+          cy.get(selector)
+            .eq(1)
+            .should('have.attr', 'rel', 'dns-prefetch');
+        });
+      });
+    });
+
+    describeForEuOnly('Consent Banners', () => {
+      it('have correct translations', () => {
+        cy.hasConsentBannerTranslations(service);
+      });
+    });
+
+    describe('Header Tests', () => {
+      it('should render the BBC News branding', () => {
+        cy.get('header a').should(
+          'contain',
+          appConfig[service].serviceLocalizedName !== undefined
+            ? `${appConfig[service].product}, ${appConfig[service].serviceLocalizedName}`
+            : appConfig[service].product,
+        );
+      });
+
+      it('should have a visible banner', () => {
+        cy.get('header')
+          .should('have.lengthOf', 1)
+          .find('div[class^="Banner"]')
+          .children()
+          .should('have.lengthOf', 1)
+          .children()
+          .should('have.attr', 'href', `/${service}`)
+          .find('svg')
+          .should('be.visible');
+      });
+    });
+
+    describe('Footer Tests', () => {
+      describe('footer tests', () => {
+        it('should have a visible footer', () => {
+          cy.get('footer')
+            .should('have.length', 1)
+            .should('have.attr', 'role', 'contentinfo')
+            .find('a')
+            .should('have.attr', 'href', `/${service}`)
+            .find('svg')
+            .should('be.visible');
+        });
+      });
+
+      it('should render the BBC News branding', () => {
+        cy.get('footer a')
+          .eq(0)
+          .should(
+            'contain',
+            appConfig[service].serviceLocalizedName !== undefined
+              ? `${appConfig[service].product}, ${appConfig[service].serviceLocalizedName}`
+              : appConfig[service].product,
+          );
+      });
+
+      it('should have working links', () => {
+        cy.get('footer ul').within(() =>
+          appConfig[service].footer.links.forEach(({ href }, key) =>
+            cy.checkLinks(key, href),
+          ),
+        );
+      });
+
+      it('should contain copyright text', () => {
+        cy.get('footer p').should(
+          'contain',
+          `© ${new Date().getFullYear()} ${
+            appConfig[service].footer.copyrightText
+          }`,
+        );
+      });
+
+      it('should contain a link in the copyright text', () => {
+        cy.get('footer p')
+          .children('a')
+          .should('have.attr', 'href')
+          .and('contain', appConfig[service].footer.externalLink.href);
+      });
+    });
+  });
+};
+
+export default alwaysTests;
